Fixed copy and password strength
Made it so the UI checks for how many settinsg the user chose to strengthen their password. If all 4 checkboxes were checked, it was a very strong password, and every time they removed one, it would go down in strength.

for the copy method, I took out the alert message because it was preventing the password from being copied to clipboard.